# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Readoteca API
  version: 0.0.1
  description: OAuth-only book tracking API.
servers:
  - url: http://localhost:8080
    description: Local backend

components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id

  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
      required:
        - id
        - email
        - displayName
        - avatarUrl

    BookStatus:
      type: string
      enum:
        - want_to_read
        - reading
        - finished
        - abandoned

    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        source:
          type: string
        sourceId:
          type: string
        isbn10:
          type: string
          nullable: true
        isbn13:
          type: string
          nullable: true
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        publishedDate:
          type: string
          nullable: true
      required:
        - id
        - source
        - sourceId
        - isbn10
        - isbn13
        - title
        - authors
        - description
        - coverUrl
        - publishedDate

    CatalogBook:
      type: object
      properties:
        source:
          type: string
        sourceId:
          type: string
        isbn10:
          type: string
          nullable: true
        isbn13:
          type: string
          nullable: true
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        publishedDate:
          type: string
          nullable: true
      required:
        - source
        - sourceId
        - isbn10
        - isbn13
        - title
        - authors
        - description
        - coverUrl
        - publishedDate

    LibraryItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/BookStatus"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        notes:
          type: string
          nullable: true
        review:
          type: string
          nullable: true
        reviewPublished:
          type: boolean
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        abandonedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        book:
          $ref: "#/components/schemas/Book"
      required:
        - id
        - status
        - progress
        - rating
        - notes
        - review
        - reviewPublished
        - startedAt
        - finishedAt
        - abandonedAt
        - createdAt
        - updatedAt
        - book

    AddLibraryItemRequest:
      type: object
      properties:
        source:
          type: string
          enum:
            - google_books
        sourceId:
          type: string
      required:
        - source
        - sourceId

    PatchLibraryItemRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/BookStatus"
        progress:
          type: integer
          minimum: 0
          maximum: 100
        rating:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        notes:
          type: string
          nullable: true
        review:
          type: string
          nullable: true
        reviewPublished:
          type: boolean
      additionalProperties: false

    SearchBooksParams:
      type: object
      properties:
        q:
          type: string
          minLength: 1
      required:
        - q

paths:
  /google/login:
    get:
      operationId: googleLogin
      tags:
        - Auth
      summary: Start Google OAuth login
      parameters:
        - name: next
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirects to Google OAuth.
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string

  /google/callback:
    get:
      operationId: googleCallback
      tags:
        - Auth
      summary: Complete Google OAuth login
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Sets session cookie and redirects to the frontend callback.
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
        "400":
          description: Invalid OAuth callback.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /me:
    get:
      operationId: getMe
      tags:
        - Auth
      summary: Get the current authenticated user
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /logout:
    post:
      operationId: logout
      tags:
        - Auth
      summary: Delete the current session
      security:
        - SessionAuth: []
      responses:
        "204":
          description: Session deleted.
          headers:
            Set-Cookie:
              schema:
                type: string

  /books/search:
    get:
      operationId: searchBooks
      tags:
        - Books
      summary: Search the external book catalog
      security:
        - SessionAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Matching catalog books.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogBook"
        "400":
          description: Invalid query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /library:
    get:
      operationId: listLibrary
      tags:
        - Library
      summary: List the current user's library
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Library items.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryItem"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addLibraryItem
      tags:
        - Library
      summary: Add a catalog book to the current user's library
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLibraryItemRequest"
      responses:
        "200":
          description: Existing library item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryItem"
        "201":
          description: Created library item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryItem"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /library/{id}:
    get:
      operationId: getLibraryItem
      tags:
        - Library
      summary: Get one current-user library item
      security:
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Library item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryItem"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchLibraryItem
      tags:
        - Library
      summary: Update one current-user library item
      security:
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchLibraryItemRequest"
      responses:
        "200":
          description: Updated library item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryItem"
        "400":
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLibraryItem
      tags:
        - Library
      summary: Remove a book from the current user's library
      security:
        - SessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Removed from library.
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"